{"schema":"decodex/commit/1","summary":"Move frozen auto-center runtime","authority":"manual"}
