+++ packages/rsnap-overlay/src/overlay.rs
@@ -8,6 +8,7 @@ mod cursor_context_runtime;
 mod cursor_runtime;
 mod exit_runtime;
 mod frozen_arrow_runtime;
+mod frozen_auto_center_runtime;
 mod frozen_brush_runtime;
 mod frozen_capture_backend_adapter;
 mod frozen_capture_session_runtime;
